+++ app/Http/Controllers/Admin/HappeningController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers\Admin;
 
-use App\Events\HappeningVerified;
 use App\Events\HappeningCreated;
 use App\Events\HappeningDeleted;
 use App\Events\HappeningUpdated;
@@ -10,21 +9,19 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\StoreHappeningRequest;
 use App\Http\Requests\Admin\UpdateHappeningRequest;
 use App\Models\Happening;
-use App\Models\Institution;
 use App\Models\Resource;
+use App\Models\User;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
 use Inertia\Inertia;
 
 class HappeningController extends Controller
 {
     public function getHappenings()
     {
-        /** @var User */
-        $user = auth()->user();
-
-        $happenings = Happening::with(['resource', 'user1', 'user2'])->get()
-            ->filter(fn ($happening) => $user->can('edit', $happening->resource->institution));
+        $happenings = Happening::with(['resource', 'user1', 'user2'])->orderBy('start')->get()
+            ->filter->isViewableByUser(auth()->user());
 
         return Inertia::render('Admin/Happenings/Index', [
             'happenings' => $happenings,
@@ -33,14 +30,23 @@ class HappeningController extends Controller
 
     public function createHappening()
     {
-        return Inertia::render('Admin/Happenings/Form');
+        $resources = Resource::active()->orderBy('title')->without('closings')->get()
+            ->filter->isUserAbleToCreateHappening(auth()->user())->values();
+
+        $users = User::all()->map(function ($user) {
+            $user->permissions = $user->getPermissions(['no verifier']);
+
+            return $user;
+        });
+
+        return Inertia::render('Admin/Happenings/Form', [
+            'resources' => $resources->map->only(['id', 'title', 'institution_id', 'is_verification_required']),
+            'users' => $users->map->only(['id', 'name', 'permissions']),
+        ]);
     }
 
     public function storeHappening(StoreHappeningRequest $request)
     {
-        $resource = Resource::findOrFail($request->resource_id);
-        Institution::abortIfUnauthorized($resource->institution);
-
         $sanitized = $request->sanitized()->merge(['reserved_at' => Carbon::now()]);
         $happening = Happening::create($sanitized->toArray());
 
@@ -51,26 +57,46 @@ class HappeningController extends Controller
 
     public function editHappening(Request $request)
     {
-        $happening = Happening::with('resource')->findOrFail($request->id);
-        Institution::abortIfUnauthorized($happening->resource->institution);
+        /** @var Happening */
+        $happening = Happening::findOrFail($request->id);
+
+        $this->authorize('adminUpdate', $happening);
 
         $happening->start_date = Carbon::parse($happening->start)->format('d.m.Y');
         $happening->start_time = Carbon::parse($happening->start)->format('H:i');
         $happening->end_date = Carbon::parse($happening->end)->format('d.m.Y');
         $happening->end_time = Carbon::parse($happening->end)->format('H:i');
 
-        return Inertia::render('Admin/Happenings/Form', $happening);
+        $resources = $happening->resource->institution->resources()
+            ->active()->orderBy('title')->without('closings')->get();
+
+        $users = User::all()->map(function ($user) {
+            $user->permissions = $user->getPermissions(['no verifier']);
+
+            return $user;
+        });
+
+        return Inertia::render('Admin/Happenings/Form', [
+            'happening' => $happening->only([
+                'id',
+                'user_id_01',
+                'user_id_02',
+                'resource_id',
+                'is_verified',
+                'verifier',
+                'start_date',
+                'start_time',
+                'end_date',
+                'end_time',
+            ]),
+            'resources' => $resources->map->only(['id', 'title', 'institution_id', 'is_verification_required']),
+            'users' => $users->map->only(['id', 'name', 'permissions']),
+        ]);
     }
 
     public function updateHappening(UpdateHappeningRequest $request)
     {
-        // Check both old and new resource for authorization
         $happening = Happening::findOrFail($request->id);
-        Institution::abortIfUnauthorized($happening->resource->institution);
-
-        $resource = Resource::findOrFail($request->resource_id);
-        Institution::abortIfUnauthorized($resource->institution);
-
         $sanitized = $request->sanitized();
         $happening->update($sanitized->toArray());
 
@@ -82,8 +108,7 @@ class HappeningController extends Controller
     public function deleteHappening(Request $request)
     {
         $happening = Happening::find($request->id);
-        Institution::abortIfUnauthorized($happening->resource->institution);
-
+        $this->authorize('adminDelete', $happening);
         $happening->delete();
 
         $happening->broadcast(HappeningDeleted::class);
